feat(directory): unify node context menu across all node types
Render the shared context menu for every TreeNode (bucket, directory,
file), not only buckets. All types now expose Open / Open in new tab;
buckets keep Edit / Delete gated on modify permission. Open in new tab
opens the node route with target="_blank".

Consolidate the former bucket-only ConnectionMenu into NodeContextMenu
and route every href through buildConnectionPath (drops the duplicate
encodeURIComponent URL build).

Remove the placeholder file/directory info flow: useNodeInfoModal, the
FileInfo/DirectoryInfo modal components, and their ModalOutlet
registrations. Info is being redesigned as a details panel separately.
